backend/requirements.txt
# Core Framework
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.18

# RAG Stack
langchain>=0.2.0
langchain-community>=0.2.0
langchain-chroma>=0.1.1
chromadb>=0.4.24

# LLM & Embeddings
langchain-google-genai>=1.0.0
google-generativeai>=0.8.0

# Document Processing
pymupdf>=1.24.0
pytesseract>=0.3.10
Pillow>=10.3.0
opencv-python>=4.9.0.80

# Utilities
pydantic>=2.7.0
python-dotenv>=1.0.1
aiofiles>=23.2.1
